+++ pyliquid/callbacks.py
@@ -1,4 +1,4 @@
-from liquid.management import Wallet, Pool
+from pyliquid.liquid.management import Wallet, Pool
 
 async def request_check_balance(target_pool: Pool) -> dict:
     _wallet: Wallet = target_pool.vaul_wallet
