+++ pyliquid/liquid/exceptions.py
@@ -1,0 +1,3 @@
+from fastapi import APIRouter
+
+router = APIRouter(prefix="internal_errors", tags=['errors'])
